+++ src/ast/Node.java
@@ -1,7 +1,10 @@
 package ast;
 import visitor.*;
+
+import java.math.BigInteger;
+
 import beaver.Symbol;
 
 public abstract class Node extends Symbol {
-    public abstract Integer accept(Visitor visitor);
+    public abstract BigInteger accept(Visitor visitor);
 }
